Move old Streaming/ headers into examples/
These are outdated and mislead users about how streaming is done in JANA2
