+++ docs/_sidebar.md
@@ -11,7 +11,6 @@
   * [Monitoring](howto/monitor-debug-gui.md "JANA2: Status Debug Monitoring")
   * [New experiments](howto/use-in-new-experiment.md "JANA2: How to use in new experiments")
   * [Group Events](howto/group-events.md "JANA2: Group Events")
-  * [Streaming](howto/streaming.md "JANA2: Streaming readout")
 - [JANA1 => 2](jana1to2/jana1-to-jana2.md)  
 - [Development](development/development.md)
 - [C++ Reference](https://jeffersonlab.github.io/JANA2/refcpp/ ':target=_blank')
