+++ docs/doxygen/doxygen_index.md
@@ -16,12 +16,6 @@ This website provides documentation for JANA2 C++ API automatically generated by
 * [JFactory](class_j_factory.html): Given an event, calculate a specific result, potentially calling other JFactories recursively to obtain any prereqs
 * [JEventProcessor](class_j_event_processor.html): Run desired JFactories over the event stream, writing results to an output file or messaging consumer
 
-## Streaming Extensions
-
-* [JStreamingEventSource](class_j_streaming_event_source.html): A starting point for structured, composable streaming
-* [JTransport](struct_j_transport.html): An interface for a generic messaging transport
-* [JMessage](struct_j_message.html): An interface for a stream buffer
-
 ## User-facing utilities
 
 * [JParameter](class_j_parameter.html): Configure the behavior of components at runtime
