+++ src/examples/misc/CMakeLists.txt
@@ -2,7 +2,6 @@
 add_subdirectory(CsvWriter)
 add_subdirectory(DstExample)
 add_subdirectory(Tutorial)
-add_subdirectory(StreamingExample)
 add_subdirectory(EventGroupExample)
 #add_subdirectory(SubeventExample)
 #add_subdirectory(SubeventCUDAExample)
