+++ src/libraries/JANA/CMakeLists.txt
@@ -118,7 +118,6 @@ set(INSTALL_RPATH_USE_LINK_PATH True)
 file(GLOB jana_headers "*.h*")
 file(GLOB jana_engine_headers "Engine/*.h*")
 file(GLOB jana_services_headers "Services/*.h*")
-file(GLOB jana_streaming_headers "Streaming/*.h*")
 file(GLOB jana_utils_headers "Utils/*.h*")
 file(GLOB jana_calibs_headers "Calibrations/*.h*")
 file(GLOB jana_geometry_headers "Geometry/*.h*")
@@ -130,7 +129,6 @@ file(GLOB jana_topology_headers "Topology/*.h*")
 install(FILES ${jana_headers} DESTINATION include/JANA)
 install(FILES ${jana_engine_headers} DESTINATION include/JANA/Engine)
 install(FILES ${jana_services_headers} DESTINATION include/JANA/Services)
-install(FILES ${jana_streaming_headers} DESTINATION include/JANA/Streaming)
 install(FILES ${jana_utils_headers} DESTINATION include/JANA/Utils)
 install(FILES ${jana_calibs_headers} DESTINATION include/JANA/Calibrations)
 install(FILES ${jana_geometry_headers} DESTINATION include/JANA/Geometry)
